переименовывает функции
